move Commands runner as an internal class
