+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Requires `formal/orm:~6.0`
 - Requires `innmind/foundation:~2.1`
 - `Formal\Migrations\SQL` has been renamed to `SQL\Runner` and is now flagged as internal
+- `Formal\Migrations\Commands` has been renamed to `Commands\Runner` and is now flagged as internal
 
 ### Removed
